// src/store/main/finance/purse/types.ts
/*
 * @Author: korealu
 * @Date: 2022-02-16 16:53:07
 * @LastEditors: Please set LastEditors
 * @LastEditTime: 2022-03-16 10:08:08
 * @Description: file content
 * @FilePath: /pofi-admin/src/store/main/finance/pay/types.ts
 */
export interface IFinancePurseType {
  pursesList: any[]
  pursesCount: number
  logList: any[]
  logCount: number
  vipList: any[]
  vipCount: number
}